+++ nova/src/circuit/augmented.rs
@@ -164,10 +164,10 @@ mod tests {
 
         augmented_circuit.generate(&mut cs);
         let shape = R1csShape::from(cs.clone());
-        let k = (shape.m().next_power_of_two() as u64).trailing_zeros();
+        let k = shape.m().next_power_of_two().trailing_zeros();
         let ck = PedersenCommitment::<G1Affine>::new(k.into(), &mut rng);
-        let u_dummy = RelaxedR1csInstance::dummy(shape.l());
-        let w_dummy = RelaxedR1csWitness::dummy(shape.m_l_1(), shape.m());
+        let u_dummy = RelaxedR1csInstance::dummy(shape.l() as usize);
+        let w_dummy = RelaxedR1csWitness::dummy(shape.m_l_1() as usize, shape.m() as usize);
 
         let (x, w) = r1cs_instance_and_witness(&cs, &shape, &ck);
         let (instance, witness) = (
